refactor: 패키지명과 enum명 수정
