+++ src/main/java/janggi/rule/BoardPositionRange.java
@@ -1,15 +1,15 @@
-package janggi.setting;
+package janggi.rule;
 
 import janggi.value.Position;
 
-public enum BoardCoordinate {
+public enum BoardPositionRange {
     X_MAX(8),
     X_MIN(0),
     Y_MAX(9),
     Y_MIN(0);
     private final int value;
 
-    BoardCoordinate(int value) {
+    BoardPositionRange(int value) {
         this.value = value;
     }
 
@@ -18,10 +18,10 @@ public enum BoardCoordinate {
     }
 
     public static void validateInRange(Position position) {
-        boolean isXOutOfRange = position.x() < BoardCoordinate.X_MIN.getValue()
-                || position.x() > BoardCoordinate.X_MAX.getValue();
-        boolean isYOutOfRange = position.y() < BoardCoordinate.Y_MIN.getValue()
-                || position.y() > BoardCoordinate.Y_MAX.getValue();
+        boolean isXOutOfRange = position.x() < BoardPositionRange.X_MIN.getValue()
+                || position.x() > BoardPositionRange.X_MAX.getValue();
+        boolean isYOutOfRange = position.y() < BoardPositionRange.Y_MIN.getValue()
+                || position.y() > BoardPositionRange.Y_MAX.getValue();
         if (isXOutOfRange || isYOutOfRange) {
             throw new IllegalArgumentException("[ERROR] x좌표는 0~8, y좌표는 0~9 사이로 입력해주세요.");
         }
