+++ src/main/java/janggi/rule/PieceAssignType.java
@@ -1,4 +1,4 @@
-package janggi.setting;
+package janggi.rule;
 
 import java.util.List;
