+++ src/main/java/janggi/strategy/PoStrategy.java
@@ -2,7 +2,7 @@ package janggi.strategy;
 
 import janggi.piece.Piece;
 import janggi.piece.PieceType;
-import janggi.setting.GungSung;
+import janggi.rule.GungSungRange;
 import janggi.value.Direction;
 import janggi.value.Path;
 import janggi.value.Position;
@@ -67,7 +67,7 @@ public class PoStrategy implements MoveStrategy {
     }
 
     private boolean isPathInGungsung(Position start, Position destination) {
-        return GungSung.isInAnyGungSung(start) && GungSung.isInAnyGungSung(destination);
+        return GungSungRange.isInAnyGungSung(start) && GungSungRange.isInAnyGungSung(destination);
     }
 
     private boolean isPossibleDirection(Direction direction, boolean isPathInGungsung) {
