+++ src/main/java/janggi/view/GameInputOutput.java
@@ -3,8 +3,8 @@ package janggi.view;
 import janggi.game.GameInformation;
 import janggi.game.JanggiBoard;
 import janggi.game.MovePieceCommand;
-import janggi.setting.CampType;
-import janggi.setting.PieceAssignType;
+import janggi.rule.CampType;
+import janggi.rule.PieceAssignType;
 import janggi.value.Position;
 import janggi.view.answer.GameMenuAnswer;
 import janggi.view.answer.TurnMenuAnswer;
