+++ src/test/java/janggi/JanggiBoardTest.java
@@ -10,8 +10,8 @@ import janggi.game.JanggiBoard;
 import janggi.game.MovePieceCommand;
 import janggi.piece.Piece;
 import janggi.piece.PieceType;
-import janggi.setting.CampType;
-import janggi.setting.PieceAssignType;
+import janggi.rule.CampType;
+import janggi.rule.PieceAssignType;
 import janggi.value.Position;
 import java.util.List;
 import java.util.stream.Stream;
