+++ src/test/java/janggi/piece/PiecesTest.java
@@ -3,7 +3,7 @@ package janggi.piece;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import janggi.setting.CampType;
+import janggi.rule.CampType;
 import janggi.value.Position;
 import java.util.List;
 import java.util.Optional;
